feta:加了很多功能

+++ internal/middleware/jwt.go
@@ -12,7 +12,7 @@ import (
 )
 
 type Claims struct {
-	UserID   uint   `json:"user_id"`
+	UserID   int64  `json:"user_id"`
 	Username string `json:"username"`
 	jwt.RegisteredClaims
 }
